# api/openapi.yaml
openapi: 3.1.0
info:
  title: Subinc Cost Management Microservice API
  version: "1.0"
  description: Production-grade API for multi-tenant SaaS cloud cost management. RBAC/ABAC endpoints included.
servers:
  - url: /api
paths:
  /users/{id}/roles:
    post:
      summary: Assign role to user
      tags: [Users, RBAC]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
      responses:
        "200":
          description: Roles updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      type: string
  /users/{id}/roles/{role}:
    delete:
      summary: Remove role from user
      tags: [Users, RBAC]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: role
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Roles updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      type: string
  /users/{id}/attributes:
    post:
      summary: Set user attribute
      tags: [Users, ABAC]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key:
                  type: string
                value:
                  type: string
      responses:
        "200":
          description: Attributes updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  attributes:
                    type: object
                    additionalProperties:
                      type: string
  /users/{id}/attributes/{key}:
    delete:
      summary: Remove user attribute
      tags: [Users, ABAC]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attributes updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  attributes:
                    type: object
                    additionalProperties:
                      type: string
...